feat: getConversionData function moved to utilities file
