+++ src/integrations/GoogleAds/utils.js
@@ -1,11 +1,15 @@
+/**
+ * if trackEvents are not enabled then do not send events (return false)
+ * if trackEvents are enabled but event filtering is not enabled then send all events
+ * if trackEvents are enabled and event filtering is also enabled then check if event is configured to event list or not
+ * if non of above is true then do not send event at all (return false)
+ * @param {*} eventName
+ * @param {*} trackEvents
+ * @param {*} enableFiltering
+ * @param {*} events
+ * @returns
+ */
 function shouldSendEvent(eventName, trackEvents, enableFiltering, events) {
-  /**
-   * if trackEvents are not enabled then do not send events (return false)
-   * if trackEvents are enabled but event filtering is not enabled then send all events
-   * if trackEvents are enabled and event filtering is also enabled then check if event is configured to event list or not
-   * if non of above is true then do not send event at all (return false)
-   */
-
   if (!trackEvents) {
     return false;
   }
@@ -27,4 +31,30 @@ function shouldSendEvent(eventName, trackEvents, enableFiltering, events) {
   return false;
 }
 
-export default shouldSendEvent;
+/**
+ * returns conversionData object (conversionLabel and eventName)
+ * if eventName is not present but defaultPageConversion is configured then keeping default eventName as Viewed a Page
+ * @param {*} eventTypeConversions
+ * @param {*} eventName
+ * @returns
+ */
+function getConversionData(eventTypeConversions, eventName) {
+  const conversionData = {};
+  if (eventTypeConversions) {
+    if (eventName) {
+      eventTypeConversions.forEach((eventTypeConversion) => {
+        if (eventTypeConversion.name.toLowerCase() === eventName.toLowerCase()) {
+          // rudderElement["message"]["name"]
+          conversionData.conversionLabel = eventTypeConversion.conversionLabel;
+          conversionData.eventName = eventTypeConversion.name;
+        }
+      });
+    } else if (this.defaultPageConversion) {
+      conversionData.conversionLabel = this.defaultPageConversion;
+      conversionData.eventName = 'Viewed a Page';
+    }
+  }
+  return conversionData;
+}
+
+export { shouldSendEvent, getConversionData };
